+++ src/Resources/contao/languages/de/tl_block_module.php
@@ -5,18 +5,20 @@ $lang = &$GLOBALS['TL_LANG']['tl_block_module'];
 /**
  * Fields
  */
-$lang['overscrollLayoutSections'][0]    = 'Anzuzeigende Layoutbereiche';
-$lang['overscrollLayoutSections'][1]    = 'Wählen Sie hier die Layoutbereiche, die im Overscroll-Bereich zu sehen sein sollen.';
-$lang['overscrollHeight'][0]            = 'Höhe des Overscroll-Containers';
-$lang['overscrollHeight'][1]            = 'Geben Sie hier die Höhe des Overscroll-Containers an.';
-$lang['overscrollCaption'][0]           = 'Text';
-$lang['overscrollCaption'][1]           = 'Geben Sie hier den Text im Overscroll-Container an.';
-$lang['overscrollLineColor'][0]         = 'Hintergrundfarbe des Fortschrittsbalkens';
-$lang['overscrollLineColor'][1]         = 'Geben Sie hier die Farbe des Fortschrittsbalkens ein.';
-$lang['overscrollLineColorFinished'][0] = 'Linienfarbe des Fortschrittsbalkens';
-$lang['overscrollLineColorFinished'][1] = 'Geben Sie hier die Farbe des Fortschrittsbalkens ein.';
-$lang['overscrollJumpTo'][0]            = 'Weiterleitungsseite';
-$lang['overscrollJumpTo'][1]            = 'Wählen Sie hier die Seite aus, zu der weitergeleitet werden soll, wenn der Nutzer über den Overscroll-Bereich gescrollt hat. Hinweis: Sie können diese Seite auch dynamisch über einen Hook anpassen.';
+$lang['overscrollLayoutSections'][0]       = 'Anzuzeigende Layoutbereiche';
+$lang['overscrollLayoutSections'][1]       = 'Wählen Sie hier die Layoutbereiche, die im Overscroll-Bereich zu sehen sein sollen.';
+$lang['overscrollHiddenLayoutSections'][0] = 'Zu versteckende Layoutbereiche';
+$lang['overscrollHiddenLayoutSections'][1] = 'Wählen Sie hier die Layoutbereiche, die bei Seiten mit Overscroll-Bereich versteckt werden sollen.';
+$lang['overscrollHeight'][0]               = 'Höhe freien Fläche unterhalb Overscroll-Containers';
+$lang['overscrollHeight'][1]               = 'Geben Sie hier die Höhe der freien Fläche unterhalb des Overscroll-Containers an.';
+$lang['overscrollCaption'][0]              = 'Text';
+$lang['overscrollCaption'][1]              = 'Geben Sie hier den Text im Overscroll-Container an.';
+$lang['overscrollLineColor'][0]            = 'Hintergrundfarbe des Fortschrittsbalkens';
+$lang['overscrollLineColor'][1]            = 'Geben Sie hier die Farbe des Fortschrittsbalkens ein.';
+$lang['overscrollLineColorFinished'][0]    = 'Linienfarbe des Fortschrittsbalkens';
+$lang['overscrollLineColorFinished'][1]    = 'Geben Sie hier die Farbe des Fortschrittsbalkens ein.';
+$lang['overscrollJumpTo'][0]               = 'Weiterleitungsseite';
+$lang['overscrollJumpTo'][1]               = 'Wählen Sie hier die Seite aus, zu der weitergeleitet werden soll, wenn der Nutzer über den Overscroll-Bereich gescrollt hat. Hinweis: Sie können diese Seite auch dynamisch über einen Hook anpassen.';
 
 /**
  * Legends