+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.0.4] - 2017-10-20
+
+### Added
+- `tl_block_module.overscrollHiddenLayoutSections` to hide layout areas on pages with overscrolling that should be visible inside overscrolling area, but not within page itself (e.g. frontpage stages)
+
+### Changed
+- `tl_block_module.overscrollHeight` is now the free gap after the overscroll container
+- full `progress-circle` container is now clickable
+
 ## [1.0.3] - 2017-10-17
 
 ### Fixed